♻️ 拆分 AIChatInput 为独立模块（content/extensions/keyboard/types/useInputDraft）
